RDBC-767 Separated cyclic deps
