+++ ravendb/tests/jvm_migrated_tests/client_tests/documents_tests/operations_tests/configuration/test_client_configuration.py
@@ -1,5 +1,5 @@
-from ravendb.documents.operations.configuration import (
-    ClientConfiguration,
+from ravendb.documents.operations.configuration.definitions import ClientConfiguration
+from ravendb.documents.operations.configuration.operations import (
     PutServerWideClientConfigurationOperation,
     GetServerWideClientConfigurationOperation,
     GetClientConfigurationOperation,
